Avoid wrapping attribute values for known types
If an attribute has known type, we should return the value without wrapping it.
It happens that when parsing the minSdkVersion attribute (<uses-sdk />) of the package `com.xunmeng.pinduoduo`, the condition `i == idAttribute` is valid, which causes invalid return value of `getAttrValue`.
This commit fix the issue https://github.com/JingMatrix/LSPatch/issues/24